requirements.txt
asgiref==3.2.5
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-guardian==2.2.0
django-js-asset==1.2.2
django-quill-editor==0.1.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.1.2
protobuf==3.12.4
PyJWT==1.7.1
pytz==2019.3
qrcode==6.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
